Move createPlaywrightMsTestTestData() to the same test
